requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.1.1
django-bootstrap-datepicker-plus==3.0.5
gunicorn==20.1.0
joblib==0.16.0
numpy==1.19.1
pandas==1.1.1
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
threadpoolctl==2.1.0
waitress==2.0.0
whitenoise==5.3.0
